Added bot location front line

+++ src/main/sensor/response/CalibrationResponse.kt
@@ -2,4 +2,4 @@ package main.sensor.response
 
 data class CalibrationResponse(val data: CalibrationResponseData) : BaseResponse(null, null)
 data class CalibrationResponseData(val pointBoard: Point, val pointImage: Point)
-data class Point(val x: Double, val y: Double)
+data class Point(val x: Float, val y: Float)
